BoxOutline rewritted to ts
