--- src/components/BoxOutline/BoxOutline.js
@@ -1,107 +1,0 @@
-import Icon from '@components/Icon'
-import { useStyles } from '@helpers/hooks/useStyles'
-import PropTypes from 'prop-types'
-import React from 'react'
-
-import styles from './BoxOutline.module.scss'
-
-/**
- * Box outline - stateless presentational component
- * @param {object} children - children
- * @param {string|array} className - list of custom classes out of component
- * @param {string} variant - box variant
- * @param {boolean} isSelected - bool to decide if box was selected or not
- * @param {func} onClickHandler - callback to invoke when box was clicked
- * @param {func} onDoubleClickHandler - callback to invoke when box was clicked double times
- * @param {string} padding - deafult is set. Set 'none' to turn off
- * @param {boolean} disableHover - disable hover events only
- * @param {boolean} disabled - grayed contend and disable all events
- * @param {boolean} noCheckmark - selected without green checkmark on top right corner
- * @param {func} onMouseOver - onMouseOver
- * @param {func} onMouseLeave - onMouseLeave
- * @return {object} An object of children element
- */
-const BoxOutline = ({
-  children,
-  className,
-  variant,
-  isSelected,
-  onClickHandler,
-  onDoubleClickHandler,
-  padding,
-  disableHover,
-  disabled,
-  noCheckmark,
-  onMouseOver,
-  onMouseLeave
-}) => {
-  const elementClasses = useStyles(
-    {
-      [styles['box-outline']]: true,
-      [styles[`box-outline__${variant}`]]: variant,
-      [styles[`box-outline__${variant}--hover`]]: !disableHover,
-      [styles['box-outline__is-selected']]: isSelected,
-      [styles['box-outline__is-selected--hover']]: isSelected && !disableHover,
-      [styles['box-outline__no-padding']]: padding === 'none',
-      [styles['box-outline__padding-large']]: padding === 'large',
-      [styles['box-outline__disabled']]: disabled
-    },
-    className
-  )
-
-  const checkmarkStatusClasses = useStyles({
-    [styles['box-outline__checkmark-status']]: true,
-    [styles['box-outline__checkmark-status--is-selected']]: isSelected
-  })
-
-  return (
-    <div
-      data-testid='box-outline'
-      className={elementClasses}
-      onClick={onClickHandler}
-      onDoubleClick={onDoubleClickHandler}
-      onMouseOver={onMouseOver}
-      onMouseLeave={onMouseLeave}
-    >
-      {isSelected && !noCheckmark && (
-        <span className={checkmarkStatusClasses}>
-          <Icon icon='icon-ok' color='white' />
-        </span>
-      )}
-      {children}
-    </div>
-  )
-}
-
-BoxOutline.propTypes = {
-  children: PropTypes.node.isRequired,
-  className: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
-  variant: PropTypes.oneOf(['default', 'background']),
-  isSelected: PropTypes.bool,
-  onClickHandler: PropTypes.func,
-  onDoubleClickHandler: PropTypes.func,
-  padding: PropTypes.oneOf(['none', 'large']),
-  disableHover: PropTypes.bool,
-  disabled: PropTypes.bool,
-  noCheckmark: PropTypes.bool,
-  onMouseOver: PropTypes.func,
-  onMouseLeave: PropTypes.func
-}
-
-BoxOutline.defaultProps = {
-  className: '',
-  variant: 'default',
-  isSelected: false,
-  onClick: () => null,
-  onDoubleClick: () => null,
-  padding: null,
-  disableHover: false,
-  disabled: false,
-  noCheckmark: false,
-  onMouseOver: () => null,
-  onMouseLeave: () => null
-}
-
-BoxOutline.displayName = 'BoxOutline'
-
-export default BoxOutline
